insert and update pictures

+++ app/Http/Controllers/PostsController.php
@@ -3,24 +3,22 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
+use Intervention\Image\Facades\Image;
 use App\Models\Post;
 use DB;
 
 class PostsController extends Controller
 {
-
-
-     /**
+    /**
      * Create a new controller instance.
      *
      * @return void
      */
     public function __construct()
     {
-        $this->middleware('auth',['except'=>['index','show']]);
+        $this->middleware('auth', ['except' => ['index', 'show']]);
     }
-
-
 
     /**
      * Display a listing of the resource.
@@ -29,16 +27,14 @@ class PostsController extends Controller
      */
     public function index()
     {
-        // $posts = Post::all();
-        // return $posts = Post::where('title','Post Two')->get();
-        // $posts = DB::select('Select * From posts');
-        //  $posts = Post::orderBy('title','desc')->take(1)->get();
-
-
-        // $posts = Post::orderBy('title','desc')->get();
+        //$posts = Post::all();
+        //return Post::where('title', 'Post Two')->get();
+        //$posts = DB::select('SELECT * FROM posts');
+        //$posts = Post::orderBy('title','desc')->take(1)->get();
+        //$posts = Post::orderBy('title','desc')->get();
 
         $posts = Post::orderBy('created_at','desc')->paginate(10);
-        return view('posts.index')->with('posts',$posts);
+        return view('posts.index')->with('posts', $posts);
     }
 
     /**
@@ -49,7 +45,6 @@ class PostsController extends Controller
     public function create()
     {
         return view('posts.create');
-
     }
 
     /**
@@ -60,19 +55,44 @@ class PostsController extends Controller
      */
     public function store(Request $request)
     {
-       $this->validate($request,[
-        'title'=>'required',
-        'body'=>'required'
-       ]);
-       //Create post
-       $post = new Post;
+        $this->validate($request, [
+            'title' => 'required',
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
+        ]);
 
-       $post->title = $request->input('title');
-       $post->body = $request->input('body');
-       $post->user_id = auth()->user()->id;
-       $post->save();
+        // Handle File Upload
+        if($request->hasFile('cover_image')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+		
+	    // // make thumbnails
+	    // $thumbStore = 'thumb.'.$filename.'_'.time().'.'.$extension;
+        //     $thumb = Image::make($request->file('cover_image')->getRealPath());
+        //     $thumb->resize(80, 80);
+        //     $thumb->save('storage/cover_images/'.$thumbStore);
+		
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
-       return redirect('/posts')->with('success','Post Created');
+        // Create Post
+        $post = new Post;
+        $post->title = $request->input('title');
+        $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
+        $post->save();
+
+        return redirect('/posts')->with('success', 'Post Created');
     }
 
     /**
@@ -84,7 +104,7 @@ class PostsController extends Controller
     public function show($id)
     {
         $post = Post::find($id);
-        return view('posts.show')->with('post',$post);
+        return view('posts.show')->with('post', $post);
     }
 
     /**
@@ -95,14 +115,19 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-         $post = Post::find($id);
-
-        //check for correct user
-        if(auth()->user()->id !== $post->user_id){
-            return redirect('/posts')->with('error','Unauthorized Page');
+        $post = Post::find($id);
+        
+        //Check if post exists before deleting
+        if (!isset($post)){
+            return redirect('/posts')->with('error', 'No Post Found');
         }
 
-         return view('posts.edit')->with('post',$post);
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        return view('posts.edit')->with('post', $post);
     }
 
     /**
@@ -114,18 +139,43 @@ class PostsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request,[
-            'title'=>'required',
-            'body'=>'required'
-           ]);
-           //Create post
-           $post = Post::find($id);
-    
-           $post->title = $request->input('title');
-           $post->body = $request->input('body');
-           $post->save();
-    
-           return redirect('/posts')->with('success','Post Updated');
+        $this->validate($request, [
+            'title' => 'required',
+            'body' => 'required',
+        ]);
+		$post = Post::find($id);
+         // Handle File Upload
+        if($request->hasFile('cover_image')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+            // Delete file if exists
+            Storage::delete('public/cover_images/'.$post->cover_image);
+		
+	//    Make thumbnails
+	    // $thumbStore = 'thumb.'.$filename.'_'.time().'.'.$extension;
+        //     $thumb = Image::make($request->file('cover_image')->getRealPath());
+        //     $thumb->resize(80, 80);
+        //     $thumb->save('storage/cover_images/'.$thumbStore);
+        
+        }
+
+        // Update Post
+        $post->title = $request->input('title');
+        $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
+        $post->save();
+
+        return redirect('/posts')->with('success', 'Post Updated');
     }
 
     /**
@@ -137,13 +187,23 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-
-        if(auth()->user()->id !== $post->user_id){
-            return redirect('/posts')->with('error','Unauthorized Page');
+        
+        //Check if post exists before deleting
+        if (!isset($post)){
+            return redirect('/posts')->with('error', 'No Post Found');
         }
 
-        $post->delete();
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
 
-        return redirect('/posts')->with('success','Post Removed');
+        if($post->cover_image != 'noimage.jpg'){
+            // Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+        
+        $post->delete();
+        return redirect('/posts')->with('success', 'Post Removed');
     }
 }
